added  rosserial_xbee
Rewrote rosserial_python's parsing code to use non blocking io.  This allows it to handle the case when it reads a corrupted data_length and would hang as it tries to read an impossibly large number of bytes.  Now, if characters stop flowing after 0.01 seconds while the message is started, all of the message data is thrown away and the parsing state machine is restarted.  This fixed alot of bugs with Xbees.

Modified ros_lib.cpp so that it computes the checksum for a packet and then continously writes the packet to the output.

sending Empty messages seem to be working.  I dont know why.